# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FunnyActivities API",
    "description": "A comprehensive API for managing materials in the FunnyActivities system. Supports CRUD operations, bulk operations, advanced filtering, and role-based access control.",
    "contact": {
      "name": "FunnyActivities Support",
      "email": "support@funnyactivities.com"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/steps/{id}": {
      "get": {
        "tags": [
          "Step"
        ],
        "summary": " (Requires: CanViewStep)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Step"
        ],
        "summary": " (Requires: CanUpdateStep)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.UpdateStepRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.UpdateStepRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.UpdateStepRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "delete": {
        "tags": [
          "Step"
        ],
        "summary": " (Requires: CanDeleteStep)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/steps/by-activity/{activityId}": {
      "get": {
        "tags": [
          "Step"
        ],
        "summary": " (Requires: CanViewStep)",
        "parameters": [
          {
            "name": "activityId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/steps": {
      "post": {
        "tags": [
          "Step"
        ],
        "summary": " (Requires: CanCreateStep)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.CreateStepRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.CreateStepRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.CreateStepRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnyActivities.Application.DTOs.ActivityManagement.StepDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FunnyActivities.Application.DTOs.ActivityManagement.CreateStepRequest": {
        "required": [
          "activityId",
          "description",
          "order"
        ],
        "type": "object",
        "properties": {
          "activityId": {
            "type": "string",
            "description": "Gets or sets the activity ID.",
            "format": "uuid"
          },
          "order": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "description": "Gets or sets the order of the step in the activity.",
            "format": "int32"
          },
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the description of the step."
          }
        },
        "additionalProperties": false,
        "description": "Request model for creating a new step."
      },
      "FunnyActivities.Application.DTOs.ActivityManagement.StepDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets or sets the unique identifier of the step.",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "description": "Gets or sets the activity ID.",
            "format": "uuid"
          },
          "activityName": {
            "type": "string",
            "description": "Gets or sets the activity name.",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "description": "Gets or sets the order of the step in the activity.",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "description": "Gets or sets the description of the step.",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "description": "Gets or sets the date and time when the step was created.",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Gets or sets the date and time when the step was last updated.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Data transfer object for step information."
      },
...
      "FunnyActivities.Application.DTOs.ActivityManagement.UpdateStepRequest": {
        "required": [
          "description",
          "order"
        ],
        "type": "object",
        "properties": {
          "order": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "description": "Gets or sets the order of the step in the activity.",
            "format": "int32"
          },
          "description": {
            "maxLength": 1000,
            "minLength": 1,
            "type": "string",
            "description": "Gets or sets the description of the step."
          }
        },
        "additionalProperties": false,
        "description": "Request model for updating an existing step."
      },
...
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Please enter JWT with Bearer into field. Example: Bearer {token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}